move level to its own file
